requirements.txt
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
tensorflow==2.12.0
pillow==9.5.0
fastapi==0.95.2
uvicorn==0.22.0
